move base.{EventListener,Logger} to pebble.{EventListener,Logger}
Now that `base.{EventListener,Logger} are only used by the `pebble`
package we can replace the type aliases with the actual types.
