--- internal/base/logger.go
@@ -1,31 +1,0 @@
-// Copyright 2011 The LevelDB-Go and Pebble Authors. All rights reserved. Use
-// of this source code is governed by a BSD-style license that can be found in
-// the LICENSE file.
-
-package base
-
-import (
-	"fmt"
-	"log"
-	"os"
-)
-
-// Logger defines an interface for writing log messages.
-type Logger interface {
-	Infof(format string, args ...interface{})
-	Fatalf(format string, args ...interface{})
-}
-
-// DefaultLogger logs to the Go stdlib logs.
-type DefaultLogger struct{}
-
-// Infof implements the Logger.Infof interface.
-func (DefaultLogger) Infof(format string, args ...interface{}) {
-	_ = log.Output(2, fmt.Sprintf(format, args...))
-}
-
-// Fatalf implements the Logger.Fatalf interface.
-func (DefaultLogger) Fatalf(format string, args ...interface{}) {
-	_ = log.Output(2, fmt.Sprintf(format, args...))
-	os.Exit(1)
-}
