+++ logger.go
@@ -5,8 +5,27 @@
 package pebble
 
 import (
-	"github.com/cockroachdb/pebble/internal/base"
+	"fmt"
+	"log"
+	"os"
 )
 
-// Logger exports the base.Logger type.
-type Logger = base.Logger
+// Logger defines an interface for writing log messages.
+type Logger interface {
+	Infof(format string, args ...interface{})
+	Fatalf(format string, args ...interface{})
+}
+
+// DefaultLogger logs to the Go stdlib logs.
+type DefaultLogger struct{}
+
+// Infof implements the Logger.Infof interface.
+func (DefaultLogger) Infof(format string, args ...interface{}) {
+	_ = log.Output(2, fmt.Sprintf(format, args...))
+}
+
+// Fatalf implements the Logger.Fatalf interface.
+func (DefaultLogger) Fatalf(format string, args ...interface{}) {
+	_ = log.Output(2, fmt.Sprintf(format, args...))
+	os.Exit(1)
+}
